v2-optimize/requirements.txt
# Bradford Bulls v2 — Smart Frame Extraction
opencv-python-headless>=4.8
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
tqdm>=4.65
scikit-learn>=1.3
scikit-image>=0.21
imagehash>=4.3
Pillow>=10.0
ultralytics>=8.3
torch>=2.0
roboflow>=1.1
